changed assigned and grantlog to be the same

+++ genghis.py
@@ -24,17 +24,22 @@ def forbid(user, ui):
 			da = [l.rstrip().split(',') for l in fa.readlines()]
 
 		remove = []
+		effectedUsers = []
 		for i,l in enumerate(da):
-			u,t,b = l
+			u1, u2,t,b = l
 			t = t.rstrip()
-			if u == ui[1] and t == ui[2]:
-				print('ERROR: '+u+' is assigned to '+t)
-				print('Inserting ({U},{T}) into forbidden may result in disruption of operations.'.format(U = u, T = t))
-				confirm = input('Are you sure you want to proceed? (y/n)>')
-				confirm = confirm == 'y'
-				if not confirm:
-					return 0
+			if u2 == ui[1] and t == ui[2]:
+				print('ERROR: '+u2+' is assigned to '+t+' by '+u1)
 				remove.append(i)
+				effectedUsers.append(u1)
+
+		if len(remove) > 0:
+			print('Inserting ({U},{T}) into forbidden may result in disruption of operations.'.format(U = u2, T = t))
+			confirm = input('Are you sure you want to proceed? (y/n)>')
+			confirm = confirm == 'y'
+			if not confirm:
+				return 0
+				
 		[da.pop(i) for i in remove]
 
 		with open('assigned.csv', 'w') as fw:
@@ -47,63 +52,19 @@ def forbid(user, ui):
 	with open('forbidden.csv', 'a') as f:
 		f.write(','.join(ui[1:])+'\n')
 
-	message = '{date} -- You have been forbidden access to {table} by admin'.format(
+	message1 = '{date} -- You have been forbidden access to {table} by admin'.format(
 				date = datetime.datetime.now().strftime("%Y-%m-%d %H:%M"),
 				table = ui[2])
 
-	leaveMessage(ui[1], message)
+	leaveMessage(ui[1], message1)
 
-	return 1
-
-def assign(user, ui):
-	if user != 'admin':
-		print("ERROR: Only admin has access to forbid function")
-		return 0
-	if len(ui) < 3:
-		print("USAGE: assign <user> <table>")
-		return 0
-	try:
-		with open('assigned.csv', 'r') as f:
-			d =[l.rstrip().split(',') for l in f.readlines()]
-
-		for u,t,b in d:
-			t = t.rstrip()
-			if u == ui[1] and t == ui[2]:
-				print('ERROR: '+u+' is already assigned to '+t)
-				return 0
-
-		with open('forbidden.csv', 'r') as fa:
-			da = [l.rstrip().split(',') for l in fa.readlines()]
-
-		remove = []
-		for i,l in enumerate(da):
-			u,t = l
-			t = t.rstrip()
-			if u == ui[1] and t == ui[2]:
-				print('ERROR: '+u+' is forbidden from '+t)
-				print('Inserting ({U},{T}) into assigned may result in security issues'.format(U = u, T = t))
-				confirm = input('Are you sure you want to proceed? (y/n)>')
-				confirm = confirm == 'y'
-				if not confirm:
-					return
-				remove.append(i)
-		[da.pop(i) for i in remove]
-
-		with open('forbidden.csv', 'w') as fw:
-			for l in da:
-				fw.write(','.join(l)+'\n')
-					
-	except FileNotFoundError:
-		pass
-
-	with open('assigned.csv', 'a') as f:
-		f.write(','.join(ui[1:])+',1\n')
-
-	message = '{date} -- You have been assigned access to {table} by admin'.format(
+	message2 = '{date} -- {u2} has been forbidden access to {table} by admin'.format(
+				u2 = ui[1],
 				date = datetime.datetime.now().strftime("%Y-%m-%d %H:%M"),
 				table = ui[2])
 
-	leaveMessage(ui[1], message)
+	for _user in effectedUsers:
+		leaveMessage(_user, message2)
 
 	return 1
 
@@ -111,19 +72,18 @@ def create(user, ui):
 	if len(ui) != 2:
 		print('USAGE: CREATE <tablename>')
 		return 0
-	with open('grantLog.csv','a') as f:
-		f.write(','.join(['admin',user,ui[1],'1','+\n']))
+	if user == '':
+		print('ERROR: Must be logged in to perform this action')
+		return 0
 
-	with open('assigned.csv', 'a') as f:
-		f.write('admin,' + ui[1] + ',' +'1' + '\n')
-		if user != '':
-			f.write(user + ',' + ui[1] + ',' +'1' + '\n')
+	with open('assigned.csv','a') as f:
+		f.write(','.join(['admin',user,ui[1],'1']) +'\n')
 
 	return 1
 
 def grant(user, ui):
-	if len(ui) != 5:
-		print('USAGE: GRANT <username> <table> <grantOption> <action>')
+	if len(ui) != 4:
+		print('USAGE: GRANT <username> <table> <grantOption>')
 		return 0
 	if user == '':
 		print('Must be logged in to perform this action!')
@@ -146,28 +106,8 @@ def grant(user, ui):
 	if error:
 		return 0
 
-	with open('grantLog.csv', 'a') as f:
+	with open('assigned.csv', 'a') as f:
 		f.write(','.join([user] + ui[1:])+'\n')
-
-	try:
-		with open('assigned.csv', 'r') as f:
-			d = [l.rstrip().split(',') for l in f.readlines()]
-
-		found = False
-		for i,l in enumerate(d):
-			user, table, b = l
-			if user == ui[1] and table == ui[2]:
-				d[i] = user,table, str(int(ui[3]) or int(b))
-				found = True
-		if not found:
-			d.append(ui[1:4])
-
-	except FileNotFoundError:
-		d = [ui[1:4]]
-
-	with open('assigned.csv', 'w') as f:
-		for l in d:
-			f.write(','.join(l)+'\n')
 
 	message = '{date} -- You have been granted access to {table} by {gu}'.format(
 				date = datetime.datetime.now().strftime("%Y-%m-%d %H:%M"),
@@ -240,11 +180,11 @@ def authenticate(user, fileName, permission):
 			l = nf.readline().rstrip().split(',')
 
 	usergraph = defaultdict(list)
-	with open('grantLog.csv', 'r') as f:
+	with open('assigned.csv', 'r') as f:
 		line = f.readline()
 		while line != '':
-			u1, u2, fn, b, p = line.rstrip().split(',')
-			if fn == fileName and (p == permission or p == '+') and not forbidden[u2][fn]:
+			u1, u2, fn, b = line.rstrip().split(',')
+			if fn == fileName and not forbidden[u2][fn]:
 				usergraph[u1].append([u2, b])
 			line = f.readline()
